+++ src/resources/views/front/werkn-backbone-bootstrap/cart.blade.php
@@ -63,8 +63,8 @@
                                     $variant = $product['variant'];
                                 @endphp
                                 <tr>
-                                    <td class="product-thumbnail"><a href="{{ route('detail', [$product['item']['category']['slug'], $product['item']['slug']]) }}">
-                                        <img src="{{ asset('img/products/' . $item_img ) }}" alt=""></a></td>
+                                    <td class="product-thumbnail" ><a href="{{ route('detail', [$product['item']['category']['slug'], $product['item']['slug']]) }}">
+                                        <img width="100" height="100" src="{{ asset('img/products/' . $item_img ) }}" alt=""></a></td>
                                     <td class="product-name">
                                         <h4><a href="{{ route('detail', [$product['item']['category']['slug'], $product['item']['slug']]) }}">{{ $product['item']['name'] }}</a></h4>
                                         <p class="mb-2">Talla: {{ $variant }}</p>