+++ src/resources/views/front/werkn-backbone-bootstrap/catalog_filter.blade.php
@@ -9,6 +9,28 @@
 @endpush
 
 @section('content')
+{{-- 
+<!-- breadcrumb-area -->
+<section class="breadcrumb-area breadcrumb-bg" data-background="{{ asset('themes/werkn-backbone-bootstrap/img/bg/breadcrumb_bg01.jpg') }}">
+    <div class="container">
+        <div class="row">
+            <div class="col-12">
+                <div class="breadcrumb-content">
+                    <h2 class="text-white">Catálogo</h2>
+                    <nav aria-label="breadcrumb">
+                        <ol class="breadcrumb">
+                            <li class="breadcrumb-item text-white"><a href="{{ route('index') }}">Inicio</a></li>
+                            <li class="breadcrumb-item active text-white" style="opacity:.7;" aria-current="page">Catálogo</li>
+                        </ol>
+                    </nav>
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
+<!-- breadcrumb-area-end -->
+--}}
+
 <!-- shop-area -->
 <section class="shop-area pt-100 pb-100">
     <div class="container">
@@ -21,6 +43,7 @@
         @else
         <div class="row">
             <div class="col-xl-3 col-lg-4">
+                
                 @include('front.theme.werkn-backbone-bootstrap.layouts.utilities._filter_sidebar')
             </div>
 
@@ -30,24 +53,14 @@
                         <div class="col-md-6">
                             <div class="shop-top-left">
                                 <ul>
-                                    <li>Mostrando {{ $products->count() }} resultados en esta selección</li>
+                                    <!--<li><a href="#"><i class="flaticon-menu"></i> FILTER</a></li>-->
+                                    <!--<li>Mostrando 1–9 de 80 resultados</li>-->
                                 </ul>
                             </div>
                         </div>
                         <div class="col-md-6">
-                            <!--
                             <div class="shop-top-right">
-                                <form action="#">
-                                    <select name="select">
-                                        <option value="">Ordenar por tendencia</option>
-                                        <option>De mayor a menor precio</option>
-                                        <option>De menor a mayor precio</option>
-                                        <option>Descuentos</option>
-                                        <option>Alfabético</option>
-                                    </select>
-                                </form>
                             </div>
-                            -->
                         </div>
                     </div>
                 </div>