+++ src/resources/views/front/werkn-backbone-bootstrap/detail.blade.php
@@ -37,7 +37,7 @@
 @endpush
 
 @section('content')
-
+<br>
 <section class="product-detail">
     <div class="container">
         <div class="row">
@@ -63,44 +63,38 @@
                 @endif
             </div>
         </div>
-
         <div class="row">
             <div class="col-md-12">
-                <ol class="breadcrumb">
-                    <li class="breadcrumb-item"><a href="{{ route('index') }}">Inicio</a></li>
-                    <li class="breadcrumb-item"><a href="{{ route('catalog.all') }}">Catálogo</a></li>
-                    <li class="breadcrumb-item"><a href="{{ route('catalog.all') }}">{{ $product->category->name ?? 'Sin Categoría' }}</a></li>
-                    <li class="breadcrumb-item active" aria-current="page">{{ $product->name }}</li>
-                </ol>
+          
             </div>
 
             <div class="col-md-6">
                 <div class="shop-details-flex-wrap d-flex">
                     <div class="shop-details-nav-wrap">
 
-                        <ul class="nav nav-tabs" id="myTab" role="tablist">
-                            <li class="nav-item" role="presentation">
-                                <a class="nav-link active" id="item-one-tab" data-toggle="tab" href="#item-one" role="tab" aria-controls="item-one" aria-selected="true"><img src="{{ asset('img/products/' . $product->image) }}" alt="" class="img-fluid"></a>
+                        <ul class=" list-group px-4" id="myTab" role="tablist">
+                            <li class="nav-item list-group-item" role="presentation">
+                                <a class="nav-link" id="item-one-tab" data-toggle="tab" href="#item-one" role="tab" aria-controls="item-one" aria-selected="true"><img width="80" height="80" src="{{ asset('img/products/' . $product->image) }}" alt="" class="img-fluid"></a>
                             </li>
                             @foreach($product->images as $image)
-                            <li class="nav-item" role="presentation">
-                                <a class="nav-link" id="item-two-{{ $image->id }}" data-toggle="tab" href="#item-{{ $image->id }}" role="tab" aria-controls="item-{{ $image->id }}" aria-selected="false"><img src="img/product/sd_nav_img02.jpg" alt="" class="img-fluid"></a>
+                            <li class="nav-item list-group-item" role="presentation">
+                                <a class="nav-link" id="item-two-{{ $image->id }}" data-toggle="tab" href="#item-{{ $image->id }}" role="tab" aria-controls="item-{{ $image->id }}" aria-selected="false"><img width="80" height="80"src="{{ asset('img/products/' . $image->image) }}" alt="" class="img-fluid"></a>
                             </li>
                             @endforeach
                         </ul>
                     </div>
 
-                    <div class="shop-details-img-wrap">
+                    <div class="shop-details-img-wrap ">
                         <div class="tab-content" id="myTabContent">
                             <div class="tab-pane fade show active" id="item-one" role="tabpanel" aria-labelledby="item-one-tab">
                                 <div class="shop-details-img">
-                                    <img src="{{ asset('img/products/' . $product->image) }}" alt="" class="img-fluid">
+                                    <img height="400" width="400" src="{{ asset('img/products/' . $product->image) }}" alt="" class="img-fluid">
                                 </div>
                             </div>
                             @foreach($product->images as $image)
                             <div class="tab-pane fade" id="item-{{ $image->id }}" role="tabpanel" aria-labelledby="item-two-{{ $image->id }}">
                                 <div class="shop-details-img">
-                                    <!--<img src="img/product/shop_details_img02.jpg" alt="" class="img-fluid">-->
+                                    <img src="{{ asset('img/products/' . $image->image) }}">
                                 </div>
                             </div>
                             @endforeach
@@ -110,10 +104,15 @@
             </div>
             <div class="col-md-6">
                 <div class="shop-details-content">
-                    <a href="{{ route('catalog', $product->category->slug) }}" class="product-cat">{{ $product->category->name ?? 'Sin Categoria' }}</a>
-                    <h3 class="title">{{ $product->name }}</h3>
+                          <ol class="breadcrumb mb-10 mx-2 mt-2">
+                    <li class="breadcrumb-item"><a href="{{ route('index') }}">Inicio</a></li>
+                    <li class="breadcrumb-item"><a href="{{ route('catalog.all') }}">Catálogo</a></li>
+                    <li class="breadcrumb-item"><a href="{{ route('catalog.all') }}">{{ $product->category->name ?? 'Sin Categoría' }}</a></li>
+                    <li class="breadcrumb-item active" aria-current="page">{{ $product->name }}</li>
+                    </ol>
+                    <h2 class="title mx-2 mt-2">{{ $product->name }}</h2>
 
-                    <div class="rating d-flex">
+                    <div class="rating d-flex mx-2 mt-2">
                         <ion-icon name="star-outline"></ion-icon>
                         <ion-icon name="star-outline"></ion-icon>
                         <ion-icon name="star-outline"></ion-icon>
@@ -121,19 +120,17 @@
                         <ion-icon name="star-outline"></ion-icon>
                     </div>
                     
-                    <p class="style-name">SKU : {{ $product->sku }}</p>
-                    <p class="style-name">Stock : {{ $product->stock }}</p>
+                    <p class="style-name mx-2 mt-2">SKU : {{ $product->sku }}</p>
+                    <p class="style-name mx-2 mt-2">Stock : {{ $product->stock }}</p>
                     @if($product->has_discount == true)
-                    <div class="price">Precio : ${{ number_format($product->discount_price, 2) }}</div>
-                    <div class="price price-discounted">${{ number_format($product->price, 2) }}</div>
+                    <div class="price mx-2 mt-2">Precio : ${{ number_format($product->discount_price, 2) }}</div>
+                    <div class="price price-discounted mx-2 mt-2">${{ number_format($product->price, 2) }}</div>
                     @else
-                    <div class="price">Precio : ${{ number_format($product->price, 2) }}</div>
+                    <div class="price mx-2 mt-2">Precio : ${{ number_format($product->price, 2) }}</div>
                     @endif
 
                     
-                    <div class="product-details-info">
-                        <span><a href="javascript:void(0)" data-toggle="modal" data-target="#sizesModal"><i class="fas fa-shoe-prints"></i> Guía de tallas</a></span>
-
+                    <div class="product-details-info mx-2 mt-2">
                         @if($product->has_variants == true)
                             @if($product->variants->count() != 0)
                             <div class="sidebar-product-size mb-30">
@@ -143,9 +140,9 @@
                                         @foreach($product->variants as $variant)
                                             <li>
                                                 @if($variant->pivot->stock <= 0)
-                                                <div class="no-stock-variant"><span class="line"></span>{{ $variant->value }}</div>
+                                                <div class="no-stock-variant card"><span class="line"></span>{{ $variant->value }}</div>
                                                 @else
-                                                <a id="variant{{ $variant->id }}" data-value="{{ $variant->value }}" class="" href="javascript:void(0)">{{ $variant->value }}</a>
+                                                <a id="variant{{ $variant->id }}" data-value="{{ $variant->value }}" class="card px-4 pt-2 pb-2 mx-2" href="javascript:void(0)">{{ $variant->value }}</a>
                                                 @endif
                                             </li>
                                         @endforeach
@@ -155,7 +152,7 @@
                             @endif
                         @endif
                     </div>
-                    <div class="perched-info">
+                    <div class="perched-info d-flex">
                         <!--
                         <div class="cart-plus-minus">
                             <form action="#" class="num-block">
@@ -174,9 +171,9 @@
                                     <p class="no-existance-explain mb-0 mt-0"><small>Resurtiremos pronto, revisa más adelante.</small></p>
                                 </div>
                             @else
-                                <a href="#" id="cartBtn" class="btn" role="button">
-                                    <i class="fas fa-cart-plus"></i> Agregar al carrito
-                                </a>
+                           <a href="{{ route('add-cart', ['id' => $product->id, 'variant' => 'unique']) }}" id="cartBtn" class="btn btn-primary mx-2 mt-2 " role="button">
+                                <i class="fas fa-cart-plus"></i> Agregar al carrito
+                            </a>
                             @endif
                         @else
                             @if($product->stock <= 0)
@@ -185,67 +182,52 @@
                                 <p class="no-existance-explain mb-0 mt-0"><small>Resurtiremos pronto, revisa más adelante.</small></p>
                             </div>
                             @else
-                            <a href="{{ route('add-cart', ['id' => $product->id, 'variant' => 'unique']) }}" id="cartBtn" class="btn" role="button">
+                            <a href="{{ route('add-cart', ['id' => $product->id, 'variant' => 'unique']) }}" id="cartBtn" class="btn btn-primary mx-2 mt-2 " role="button">
                                 <i class="fas fa-cart-plus"></i> Agregar al carrito
                             </a>
                             @endif
                         @endif
 
-                        <div class="wishlist-compare">
-                            <ul>
-                                <li>
+                       
+                                
                                 @if(isset(Auth::user()->id) && Auth::user()->isInWishlist($product->id))
-                                    <a href="{{ route('wishlist.remove', $product->id) }}" class="wishlist-btn wishlist-btn-delete"><i class="far fa-heartbeat"></i> Quitar de tu Wishlist</a>
+                                    <a href="{{ route('wishlist.remove', $product->id) }}" class="mx-2 mt-2 btn btn-danger"><i class="far fa-heartbeat"></i> Quitar de tu Wishlist</a>
                                 @else
                                     @if(Auth::guest())
-                                    <a href="{{ route('login') }}" class="wishlist-btn d-flex align-items-center text-center p-2"><i class="far fa-heart"></i> Agregar a tu Wishlist</a>
+                                    <a href="{{ route('login') }}" class="mx-2 mt-2 btn btn-info d-flex align-items-center text-center p-2"><i class="far fa-heart"></i> Agregar a tu Wishlist</a>
                                     @else
-                                    <a href="{{ route('wishlist.add', $product->id) }}" class="wishlist-btn d-flex align-items-center text-center p-2"><i class="far fa-heart"></i> Agregar a tu Wishlist</a>
+                                    <a href="{{ route('wishlist.add', $product->id) }}" class="mx-2 mt-2 btn btn-info d-flex align-items-center text-center p-2"><i class="far fa-heart"></i> Agregar a tu Wishlist</a>
                                     @endif
                                 @endif
-                                </li>
-                            </ul>
-                        </div>
+                                
+                          
                     </div>
                 </div>
             </div>
         </div>
-
+<br>
         <div class="row">
             <div class="col-md-12">
                 <div class="product-desc-wrap">
-                    <ul class="nav nav-tabs" role="tablist">
-                        <li class="nav-item" role="presentation">
-                            <a class="nav-link active" id="description-tab" data-toggle="tab" href="#description" role="tab" aria-controls="description"
-                                aria-selected="true">Descripción y Materiales</a>
-                        </li>
-                        <li class="nav-item" role="presentation">
-                            <a class="nav-link" id="reviews-tab" data-toggle="tab" href="#reviews" role="tab" aria-controls="reviews"
-                                aria-selected="false">Reseñas ({{ $product->approved_reviews->count() ?? '0' }})</a>
-                        </li>
-                    </ul>
                     <div class="tab-content">
-                        <div class="tab-pane fade show active" id="description" role="tabpanel" aria-labelledby="description-tab">
-                            <div class="row">
-                                <div class="col-md-6">
-                                    <div class="product-desc-title mb-30">
-                                        <h4 class="title">Información adicional :</h4>
-                                    </div>
-                                    <p>{{ $product->description }}</p>
-                                </div>
-                                <div class="col-md-6">
-                                    <div class="product-desc-title mb-30">
-                                        <h4 class="title">Materiales :</h4>
-                                    </div>
-                                    <p>{{ $product->materials }}</p>
-                                </div>
-                            </div>
-                        </div>
-                        <div class="tab-pane fade" id="reviews" role="tabpanel" aria-labelledby="reviews-tab">
-                            <div class="product-desc-title mb-30">
-                                <h4 class="title">Reseñas ({{ $product->approved_reviews->count() ?? '0' }}) :</h4>
-                            </div>
-                            @if($product->approved_reviews->count() != 0)
+                        <ul class="nav nav-pills" id="myTab" role="tablist">
+                          <li class="nav-item" role="presentation">
+                            <button class="nav-link active" id="home-tab" data-bs-toggle="tab" data-bs-target="#home" type="button" role="tab" aria-controls="home" aria-selected="true"><h5 class="title">Información adicional:</h5></button>
+                          </li>
+                          <li class="nav-item" role="presentation">
+                            <button class="nav-link" id="materials-tab" data-bs-toggle="tab" data-bs-target="#materials" type="button" role="tab" aria-controls="materials" aria-selected="false"><h5 class="title">Materiales:</h5></button>
+                          </li>
+                          <li class="nav-item" role="presentation">
+                            <button class="nav-link" id="reviews-tab" data-bs-toggle="tab" data-bs-target="#reviews" type="button" role="tab" aria-controls="reviews" aria-selected="false"><h5 class="title">Reseñas ({{ $product->approved_reviews->count() ?? '0' }}) :</h5></button>
+                          </li>
+                           <li class="nav-item" role="presentation">
+                            <button class="nav-link" id="sizes-tab" data-bs-toggle="tab" data-bs-target="#sizes" type="button" role="tab" aria-controls="sizes" aria-selected="false"><h5 class="title">Guia de tallas</h5></button>
+                          </li>
+                        </ul>
+                        <div class="tab-content mt-4" id="myTabContent">
+                          <div class="tab-pane fade show active" id="home" role="tabpanel" aria-labelledby="home-tab"><p>{{ $product->description }}</p></div>
+                          <div class="tab-pane fade" id="materials" role="tabpanel" aria-labelledby="materials-tab"><p>{{ $product->materials }}</p></div>
+                          <div class="tab-pane fade" id="reviews" role="tabpanel" aria-labelledby="reviews-tab">   @if($product->approved_reviews->count() != 0)
                                 @foreach($product->approved_reviews as $review)
                                     <div class="media">
                                         <img class="mr-3 rounded-circle" src="{{ 'https://www.gravatar.com/avatar/' . md5(strtolower(trim($review->email))) . '?404' }}" width="50" alt="{{ $review->name }}">
@@ -259,38 +241,45 @@
                                     <hr>
                                 @endforeach
                             @else
-                                <p class="adara-review-title">No hay reseñas para este producto todavía. Se el primero en hablar de "{{ $product->name }}"</p>
+                                <p class="">No hay reseñas para este producto todavía. Se el primero en hablar de "{{ $product->name }}"</p>
                             @endif
-
-                            <form action="{{ route('reviews.store', $product->id) }}" method="POST" class="comment-form review-form">
+                                <form action="{{ route('reviews.store', $product->id) }}" method="POST" class="comment-form review-form">
                                 {{ csrf_field() }}
 
                                 <input type="hidden" name="product_id" value="{{ $product->id }}">
 
-                                <span>Tu reseña <span class="text-danger">*</span></span>
+                                <h5 class="mb-2">Tu reseña <span class="text-danger ">*</span></h5>
 
-                                <textarea id="review" name="review" rows="4" class="form-control" placeholder="Este producto es genial..."></textarea>
+                                <textarea id="review" name="review" rows="4" class="form-control mb-4" placeholder="Este producto es genial..."></textarea>
 
                                 @guest
                                 <div class="row">
                                     <div class="col-md-6">
-                                        <input type="text" name="name" placeholder="Tu nombre *" required="">
+                                        <h5 class="mb-2">Tu nombre <span class="text-danger ">*</span></h5>
+                                        <input type="text" name="name" placeholder="Nombre Apellido" required="">
                                     </div>
                                     <div class="col-md-6">
-                                        <input type="email" name="email" placeholder="Tu correo electrónico *" required="">
+                                        <h5 class="mb-2">Tu correo <span class="text-danger ">*</span></h5>
+                                        <input type="email" name="email" placeholder="correo@correo.com" required="">
                                     </div>
                                 </div>
                                 @else
                                 <div class="row">
                                     <div class="col-md-6">
-                                        <input type="text" name="name" placeholder="Tu nombre *" required="" value="{{ Auth::user()->name }}" readonly="">
+                                        <h5 class="mb-2">Tu nombre <span class="text-danger ">*</span></h5>
+                                        <input type="text" name="name" placeholder="Nombre Apellido" required="" value="{{ Auth::user()->name }}" readonly="">
                                     </div>
                                     <div class="col-md-6">
-                                        <input type="email" name="email" placeholder="Tu correo electrónico *" required="" value="{{ Auth::user()->email }}" readonly="">
+                                        <h5 class="mb-2">Tu correo <span class="text-danger ">*</span></h5>
+                                        <input type="email" name="email" placeholder="correo@correo.com" required="" value="{{ Auth::user()->email }}" readonly="">
                                     </div>
+
                                 </div>
-                                <small class="d-block mb-4">Tienes sesión iniciada como {{ Auth::user()->name }} 
+                                <div class="justify-content-center">
+                                      <small class="d-block mb-4">Tienes sesión iniciada como {{ Auth::user()->name }} 
                                 </small>
+                                </div>
+                              
                                 @endguest
 
                                 <!--
@@ -299,8 +288,41 @@
                                     <label for="comment-check">Save my name and email in this browser for the next time I comment.</label>
                                 </div>
                                 -->
-                                <button type="submit" class="btn">Publicar Reseña</button>
+                                <button type="submit" class="btn btn-primary mt-5">Publicar Reseña</button>
                             </form>
+                        </div>
+                             <div class="tab-pane fade show " id="sizes" role="tabpanel" aria-labelledby="reviews-tab">
+                                <table class="table table-hover table-responsive mt-2 mb-4">
+                                            <tbody>
+                                                <tr>
+                                                    <th scope="row" style="width: 150px;">MEX</th>
+                                                    <td style="text-align: center;">25.5</td>
+                                                    <td style="text-align: center;">26</td>
+                                                    <td style="text-align: center;">26.5</td>
+                                                    <td style="text-align: center;">27</td>
+                                                    <td style="text-align: center;">27.5</td>
+                                                    <td style="text-align: center;">28</td>
+                                                    <td style="text-align: center;">28.5</td>
+                                                    <td style="text-align: center;">29</td>
+                                                    <td style="text-align: center;">29.5</td>
+                                                    <td style="text-align: center;">30</td>
+                                                </tr>
+                                                <tr>
+                                                    <th scope="row" style="width: 150px;">USA</th>
+                                                    <td style="text-align: center;">-</td>
+                                                    <td style="text-align: center;">8</td>
+                                                    <td style="text-align: center;">8.5</td>
+                                                    <td style="text-align: center;">9</td>
+                                                    <td style="text-align: center;">9.5</td>
+                                                    <td style="text-align: center;">10</td>
+                                                    <td style="text-align: center;">10.5</td>
+                                                    <td style="text-align: center;">11</td>
+                                                    <td style="text-align: center;">11.5</td>
+                                                    <td style="text-align: center;">12</td>
+                                                </tr>
+                                            </tbody>
+                                        </table>
+                            </div>
                         </div>
                     </div>
                 </div>